+++ inc/object.h
@@ -9,11 +9,24 @@
 #ifndef OBJECT_H
 #define OBJECT_H
 
+#include "set.h"
 #include "types.h"
+
+enum { OBJ_YES, OBJ_NO };
+
+typedef enum _Obj_attr {
+  OBJ_IS_MOVABLE,
+  OBJ_IS_MOVED,
+  OBJ_IS_HIDDEN,
+  OBJ_CAN_ILLUMINATE,
+  OBJ_IS_ON
+} Obj_attr;
 
 #define MAX_OBJ_DESCRP 100 /*!< @brief Maximum number of characters that can have a description of an object */
 #define MAX_OBJ_NAME 50 /*!< @brief Maximum number of characters that can have a name of an object */
 #define MAX_OBJECTS 4 /*!< @brief Maximum number of objects */
+#define MAX_OBJ_ATTRS 10 /*!< @brief Maximum quantity of object attributes */
+#define MAX_OBJ_LINKS /*!< @brief Maximum number of links that an object can open */
 
 /*!
  * @brief Defines the type of object structure
@@ -46,7 +59,6 @@ void obj_destroy( Object *obj );
 */
 STATUS obj_set_name( Object *obj, const char *name );
 
-
 /**
 * @brief This fuction gets an object name
 * @author Miguel Rodríguez
@@ -60,9 +72,18 @@ const char *obj_get_name( Object *obj );
 * @brief This fuction sets the description of an object
 * @author Miguel Rodríguez
 * @param {Object*} obj - object's pointer
-* @param {char*} name - object's description
+* @param {char*} descrp - object's description
 */
 STATUS obj_set_descrp( Object *obj, const char *descrp );
+
+
+/**
+* @brief This fuction sets the long description of an object
+* @author Miguel Rodríguez
+* @param {Object*} obj - object's pointer
+* @param {char*} ldescrp - long description of the object
+*/
+STATUS obj_set_ldescrp( Object *obj, const char *ldescrp );
 
 
 /**
@@ -72,6 +93,15 @@ STATUS obj_set_descrp( Object *obj, const char *descrp );
 * @retval {char*} - Returns a pointer to the object's name
 */
 const char *obj_get_descrp( Object *obj );
+
+
+/**
+* @brief This fuction gets the long description of an object
+* @author Miguel Rodríguez
+* @param {Object*} obj - object's pointer
+* @retval {char*} - Returns the long description of the object
+*/
+const char *obj_get_ldescrp( Object *obj );
 
 
 /**
@@ -90,5 +120,11 @@ STATUS obj_set_id( Object *obj, const Id id );
 * @retval {Id} - Returns an id
 */
 const Id obj_get_id( Object *obj );
+
+bool obj_opens_link( Object *obj, Id id );
+Set* obj_get_links( Object *obj );
+STATUS obj_add_link( Object *obj, Id id );
+STATUS obj_set_attr( Object *obj, Obj_attr attr, long v );
+long obj_get_attr( Object *obj, Obj_attr attr );
 
 #endif